Added protocol and ciphersaber encryption methods
Imlimented ciphersaber 2 encryption method.
Also created methods for prepearing a message and reciving a message sent
using protocol version 0.1.
Protocol implmentation was written in a way that can read messages sent
with an older version of the protocol and potentiall send messages with an
older version of the protocol. Sending messages with an older protocol
will likely require additional user list information or an initial ping to
the user to check waht program version they are using. However, the latter
option could create a security risk. For now the only option will to send
messages using the latest protocol.