Added -user to remove users from the list
-user now removes a user from the LLL of user.

Still need to impliment a way to save the user file. Maybe after everytime
the list is modified rewrite the file? We'll see how badly that harms
performance.

Also the userfile should probably be encrypted and passworeded either with
the TauNet key or a sepoate password. Latter seems more secure.
--Recommend taunet key, but inform a different password/phrase is
moresecure (using taunet key is more likely to be bruteforced).

Log of messages recieved also needs to be implimented. Also encrypted?
Saved at all? Maybe just in memory and then discareded ? Only last 10-50 in
memory? Pi memory may be a concern when logging too many messages in
memory and not to a file. File is clunkeyer though.