Created message_list class to recored messages
message_list is a DLL that pushes (sender, reciever and the message) on to
aa stack of messages. when the 21st message is added the oldest message is
removed. print_n(int n) is a function that prints the last n messages
pushed on to the stack of messages. if n > number of messages on the stack
all messages on the stack are printed.