implimented #n command for message history
Added #n command to display the message histroy. This command only shows
the messages recieved and not the ones sent.
Listen funtion now adds recieved messages to the message list.

To add a message history that includes the user's sent messages there are
2 options:
1: Increase the message buffer size from 20 to 40(?) or more and add the
user's sent messages to the same list. This would also require some state
control to prevent writing a recieved and sent message at the same time
(although the probability of that is very, very low).
2: create a sent_hist that would act the same way as message_hist but
would require a new command or a modification of the current command to
display the different history. #@n and #n. This seems do able.

NEW IDEA:
Create #@user_name to display all messages to or from user_name. This
feature would best fit in with strategy 1 above. Increasing the message
buffer to something over 100 may be nice too if this feature is
implimented.