added the ability for the user to create a userlst
Currently when a user types a file name to laod for a file that does not
exist the system prompts the user if they would like to create one. This
guides the user through a series of steps to create the user list. A
better series of prompts when taunet starts should probably be created.

!!! The file is not actually created. A method to write the file still
needs to be created, and implimeted for +user and -user