file now written on +usr and -usr
The user list file is now written after every addition and removal of a
user. Approach is inefficient, but works well enough with a small amount
of users. As the user_list gets extrememly large the entire program will
slow down.
This problem presents a new issue with the user_list data structure in
general. A Binary search tree would be more efficient for searching the
user list. Possible implimentation in v2