Set a maximum message length and ability to resend
Set the maximum message length to be 600 characters. I also changes one of
the commands to call a function rather than processing the command in
main (I will be doing this for the rest of the commands as well).
There was an issue with recieveing very long (900> char) messages. I think
this is becase of extra things being added to the messages behind the
sceens with the socket protocol. To fix this I made the max message lenght
30% shorter than the required maximum length. If the user attempts to send
a message longer than 600 character the message gets split up and sent in
pieces.
!!! There is potentially a problem with the message relay. If the second
part of a message gets dropped, then the user doen't know if part of the
message got thorugh or not. A more verbouse message send can clear that up
though.