Added some attional exception handeling.
Had potential errors when trying to open files that did not exist
yet:
Added exception handeling to handle those errors.

message_list.print_n(-1) was improperly printing 1 message:
Added an addtional case to prevent this.
Also added excepion handeling if a wrong type is given.

After removing a user @ message would still try to send a message to that
user after the user was removed:
user_info for the previously messaged user is now set to None to prevent
this error.